fix compilation errors for non-arm64 branch

+++ core/NormalRga.cpp
@@ -1098,21 +1098,21 @@ int RgaBlit(rga_info *src, rga_info *dst, rga_info *src1) {
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 #else
         NormalRgaSetSrcVirtualInfo(&rgaReg, (unsigned long)srcBuf,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH * 5/4,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH * 5/4,
                                    srcVirW, srcVirH,
                                    RkRgaGetRgaFormat(relSrcRect.format),0);
         /* src1 */
         if (src1)
             NormalRgaSetPatVirtualInfo(&rgaReg, (unsigned long)src1Buf,
-                                       (unsigned int)src1Buf + src1VirW * src1VirH,
-                                       (unsigned int)src1Buf + src1VirW * src1VirH * 5/4,
+                                       (uintptr_t)src1Buf + src1VirW * src1VirH,
+                                       (uintptr_t)src1Buf + src1VirW * src1VirH * 5/4,
                                        src1VirW, src1VirH, &clip,
                                        RkRgaGetRgaFormat(relSrc1Rect.format),0);
         /*dst*/
         NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned long)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 
@@ -1141,9 +1141,9 @@ int RgaBlit(rga_info *src, rga_info *dst, rga_info *src1) {
                                        srcVirW, srcVirH,
                                        RkRgaGetRgaFormat(relSrcRect.format),0);
 #else
-            NormalRgaSetSrcVirtualInfo(&rgaReg, (unsigned int)srcBuf,
-                                       (unsigned int)srcBuf + srcVirW * srcVirH,
-                                       (unsigned int)srcBuf + srcVirW * srcVirH * 5/4,
+            NormalRgaSetSrcVirtualInfo(&rgaReg, (uintptr_t)srcBuf,
+                                       (uintptr_t)srcBuf + srcVirW * srcVirH,
+                                       (uintptr_t)srcBuf + srcVirW * srcVirH * 5/4,
                                        srcVirW, srcVirH,
                                        RkRgaGetRgaFormat(relSrcRect.format),0);
 #endif
@@ -1172,9 +1172,9 @@ int RgaBlit(rga_info *src, rga_info *dst, rga_info *src1) {
                                            src1VirW, src1VirH, &clip,
                                            RkRgaGetRgaFormat(relSrc1Rect.format),0);
 #else
-                NormalRgaSetPatVirtualInfo(&rgaReg, (unsigned int)src1Buf,
-                                           (unsigned int)src1Buf + src1VirW * src1VirH,
-                                           (unsigned int)src1Buf + src1VirW * src1VirH * 5/4,
+                NormalRgaSetPatVirtualInfo(&rgaReg, (uintptr_t)src1Buf,
+                                           (uintptr_t)src1Buf + src1VirW * src1VirH,
+                                           (uintptr_t)src1Buf + src1VirW * src1VirH * 5/4,
                                            src1VirW, src1VirH, &clip,
                                            RkRgaGetRgaFormat(relSrc1Rect.format),0);
 #endif
@@ -1203,9 +1203,9 @@ int RgaBlit(rga_info *src, rga_info *dst, rga_info *src1) {
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #else
-            NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned int)dstBuf,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+            NormalRgaSetDstVirtualInfo(&rgaReg, (uintptr_t)dstBuf,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #endif
@@ -1268,21 +1268,21 @@ int RgaBlit(rga_info *src, rga_info *dst, rga_info *src1) {
 
 #else
         NormalRgaSetSrcVirtualInfo(&rgaReg, srcFd != -1 ? srcFd : 0,
-                                   (unsigned int)srcBuf,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH,
+                                   (uintptr_t)srcBuf,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH,
                                    srcVirW, srcVirH,
                                    RkRgaGetRgaFormat(relSrcRect.format),0);
         /* src1 */
         if (src1)
             NormalRgaSetPatVirtualInfo(&rgaReg, src1Fd != -1 ? src1Fd : 0,
-                                       (unsigned int)src1Buf,
-                                       (unsigned int)src1Buf + src1VirW * src1VirH,
+                                       (uintptr_t)src1Buf,
+                                       (uintptr_t)src1Buf + src1VirW * src1VirH,
                                        src1VirW, src1VirH, &clip,
                                        RkRgaGetRgaFormat(relSrc1Rect.format),0);
         /*dst*/
         NormalRgaSetDstVirtualInfo(&rgaReg, dstFd != -1 ? dstFd : 0,
-                                   (unsigned int)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 
@@ -1601,9 +1601,9 @@ int RgaCollorFill(rga_info *dst) {
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 #else
         /*dst*/
-        NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned int)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+        NormalRgaSetDstVirtualInfo(&rgaReg, (uintptr_t)dstBuf,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 #endif
@@ -1631,9 +1631,9 @@ int RgaCollorFill(rga_info *dst) {
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #else
-            NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned int)dstBuf,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+            NormalRgaSetDstVirtualInfo(&rgaReg, (uintptr_t)dstBuf,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #endif
@@ -1659,8 +1659,8 @@ int RgaCollorFill(rga_info *dst) {
 #else
         /*dst*/
         NormalRgaSetDstVirtualInfo(&rgaReg, dstFd != -1 ? dstFd : 0,
-                                   (unsigned int)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
 #endif
@@ -2098,20 +2098,20 @@ int RgaCollorPalette(rga_info *src, rga_info *dst, rga_info *lut) {
                                    RkRgaGetRgaFormat(relLutRect.format),0);
 #else
         NormalRgaSetSrcVirtualInfo(&rgaReg, (unsigned long)srcBuf,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH * 5/4,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH * 5/4,
                                    srcVirW, srcVirH,
                                    RkRgaGetRgaFormat(relSrcRect.format),0);
         /*dst*/
         NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned long)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
         /*lut*/
         NormalRgaSetPatVirtualInfo(&rgaReg, (unsigned long)lutBuf,
-                                   (unsigned int)lutBuf + lutVirW * lutVirH,
-                                   (unsigned int)lutBuf + lutVirW * lutVirH * 5/4,
+                                   (uintptr_t)lutBuf + lutVirW * lutVirH,
+                                   (uintptr_t)lutBuf + lutVirW * lutVirH * 5/4,
                                    lutVirW, lutVirH, &clip,
                                    RkRgaGetRgaFormat(relLutRect.format),0);
 
@@ -2140,9 +2140,9 @@ int RgaCollorPalette(rga_info *src, rga_info *dst, rga_info *lut) {
                                        srcVirW, srcVirH,
                                        RkRgaGetRgaFormat(relSrcRect.format),0);
 #else
-            NormalRgaSetSrcVirtualInfo(&rgaReg, (unsigned int)srcBuf,
-                                       (unsigned int)srcBuf + srcVirW * srcVirH,
-                                       (unsigned int)srcBuf + srcVirW * srcVirH * 5/4,
+            NormalRgaSetSrcVirtualInfo(&rgaReg, (uintptr_t)srcBuf,
+                                       (uintptr_t)srcBuf + srcVirW * srcVirH,
+                                       (uintptr_t)srcBuf + srcVirW * srcVirH * 5/4,
                                        srcVirW, srcVirH,
                                        RkRgaGetRgaFormat(relSrcRect.format),0);
 #endif
@@ -2170,9 +2170,9 @@ int RgaCollorPalette(rga_info *src, rga_info *dst, rga_info *lut) {
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #else
-            NormalRgaSetDstVirtualInfo(&rgaReg, (unsigned int)dstBuf,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH,
-                                       (unsigned int)dstBuf + dstVirW * dstVirH * 5/4,
+            NormalRgaSetDstVirtualInfo(&rgaReg, (uintptr_t)dstBuf,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH,
+                                       (uintptr_t)dstBuf + dstVirW * dstVirH * 5/4,
                                        dstVirW, dstVirH, &clip,
                                        RkRgaGetRgaFormat(relDstRect.format),0);
 #endif
@@ -2200,9 +2200,9 @@ int RgaCollorPalette(rga_info *src, rga_info *dst, rga_info *lut) {
                                        lutVirW, lutVirH, &clip,
                                        RkRgaGetRgaFormat(relLutRect.format),0);
 #else
-            NormalRgaSetPatVirtualInfo(&rgaReg, (unsigned int)lutBuf,
-                                       (unsigned int)lutBuf + lutVirW * lutVirH,
-                                       (unsigned int)lutBuf + lutVirW * lutVirH * 5/4,
+            NormalRgaSetPatVirtualInfo(&rgaReg, (uintptr_t)lutBuf,
+                                       (uintptr_t)lutBuf + lutVirW * lutVirH,
+                                       (uintptr_t)lutBuf + lutVirW * lutVirH * 5/4,
                                        lutVirW, lutVirH, &clip,
                                        RkRgaGetRgaFormat(relLutRect.format),0);
 #endif
@@ -2262,20 +2262,20 @@ int RgaCollorPalette(rga_info *src, rga_info *dst, rga_info *lut) {
                                    RkRgaGetRgaFormat(relLutRect.format),0);
 #else
         NormalRgaSetSrcVirtualInfo(&rgaReg, srcFd != -1 ? srcFd : 0,
-                                   (unsigned int)srcBuf,
-                                   (unsigned int)srcBuf + srcVirW * srcVirH,
+                                   (uintptr_t)srcBuf,
+                                   (uintptr_t)srcBuf + srcVirW * srcVirH,
                                    srcVirW, srcVirH,
                                    RkRgaGetRgaFormat(relSrcRect.format),0);
         /*dst*/
         NormalRgaSetDstVirtualInfo(&rgaReg, dstFd != -1 ? dstFd : 0,
-                                   (unsigned int)dstBuf,
-                                   (unsigned int)dstBuf + dstVirW * dstVirH,
+                                   (uintptr_t)dstBuf,
+                                   (uintptr_t)dstBuf + dstVirW * dstVirH,
                                    dstVirW, dstVirH, &clip,
                                    RkRgaGetRgaFormat(relDstRect.format),0);
         /*lut*/
         NormalRgaSetPatVirtualInfo(&rgaReg, lutFd != -1 ? lutFd : 0,
-                                  (unsigned int)lutBuf,
-                                  (unsigned int)lutBuf + lutVirW * lutVirH,
+                                  (uintptr_t)lutBuf,
+                                  (uintptr_t)lutBuf + lutVirW * lutVirH,
                                   lutVirW, lutVirH, &clip,
                                   RkRgaGetRgaFormat(relLutRect.format),0);
